added tweet gauge component
